+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/extensions/IntelliJPlatformDependenciesHelper.kt
@@ -28,7 +28,6 @@ import org.jetbrains.intellij.platform.gradle.Constants.Configurations.Attribute
 import org.jetbrains.intellij.platform.gradle.Constants.Configurations.Dependencies
 import org.jetbrains.intellij.platform.gradle.Constants.Constraints
 import org.jetbrains.intellij.platform.gradle.Constants.Locations.GITHUB_REPOSITORY
-import org.jetbrains.intellij.platform.gradle.Constants.Sandbox
 import org.jetbrains.intellij.platform.gradle.models.*
 import org.jetbrains.intellij.platform.gradle.providers.AndroidStudioDownloadLinkValueSource
 import org.jetbrains.intellij.platform.gradle.providers.JavaRuntimeMetadataValueSource
@@ -887,10 +886,7 @@ class IntelliJPlatformDependenciesHelper(
 
         val plugin by lazy {
             val pluginPath = when {
-                artifactPath.isDirectory() -> generateSequence(artifactPath) {
-                    it.takeIf { it.resolve(Sandbox.Plugin.LIB).exists() } ?: it.listDirectoryEntries().singleOrNull()
-                }.firstOrNull { it.resolve(Sandbox.Plugin.LIB).exists() } ?: throw GradleException("Could not resolve plugin directory: '$artifactPath'")
-
+                artifactPath.isDirectory() -> artifactPath.resolvePluginPath()
                 else -> artifactPath
             }
             pluginManager.safelyCreatePlugin(pluginPath).getOrThrow()